# src/routes/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "burncpu API",
    "version": "0.0.0",
    "summary": "Public HTTP API for the burncpu social platform.",
    "description": "Machine-readable description of the burncpu API.\n\n## Authentication\n\nTwo schemes are supported:\n\n- **cookieAuth** — the `burncpu_session` HttpOnly cookie set by the magic-link\n  flow (`/auth/verify/{token}`). Required for all account-sensitive and admin\n  endpoints. State-changing requests are also CSRF-protected (same-origin).\n- **bearerAuth** — `Authorization: Bearer <token>` using an API token from\n  `POST /tokens`. Works on scoped, non-session endpoints only (e.g. read\n  timelines, create posts); it does **not** satisfy admin or 2FA-gated routes.\n\nEndpoints with no `security` are public. The `version` field is stamped from\nthe running server build.",
    "contact": { "name": "burncpu", "url": "https://burncpu.com" },
    "license": { "name": "Proprietary" }
  },
  "servers": [
    { "url": "https://burncpu.com/api/v1", "description": "Production" }
  ],
...
  "paths": {
...
    "/posts/{id}": {
      "parameters": [ { "$ref": "#/components/parameters/IdPath" } ],
      "get": { "tags": ["Posts"], "summary": "Fetch a single post", "security": [], "responses": { "200": { "description": "Post", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Post" } } } }, "404": { "$ref": "#/components/responses/NotFound" } } },
      "patch": { "tags": ["Posts"], "summary": "Edit a post (author only)", "description": "Sets `edited_at`.", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["body"], "properties": { "body": { "type": "string" }, "content_warning": { "type": "string" } } } } } }, "responses": { "200": { "description": "Updated post", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Post" } } } }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } } },
      "delete": { "tags": ["Posts"], "summary": "Soft-delete a post (author or admin)", "responses": { "204": { "description": "Deleted." }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" } } }
    },
...
    "/notifications": {
      "get": { "tags": ["Notifications"], "summary": "Notification inbox", "parameters": [ { "name": "unread_only", "in": "query", "schema": { "type": "integer", "enum": [0, 1] } }, { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Before" } ], "responses": { "200": { "description": "Notifications", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Notification" } } } } }, "401": { "$ref": "#/components/responses/Unauthorized" } } }
    },
...
    "/push/subscribe": {
      "post": { "tags": ["Push"], "summary": "Register a Web Push subscription", "security": [ { "cookieAuth": [] } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["endpoint", "keys"], "properties": { "endpoint": { "type": "string" }, "keys": { "type": "object", "properties": { "p256dh": { "type": "string" }, "auth": { "type": "string" } } } } } } } }, "responses": { "204": { "description": "Subscribed." }, "401": { "$ref": "#/components/responses/Unauthorized" } } }
    },
    "/push/unsubscribe": {
      "delete": { "tags": ["Push"], "summary": "Remove a Web Push subscription", "security": [ { "cookieAuth": [] } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["endpoint"], "properties": { "endpoint": { "type": "string" } } } } } }, "responses": { "204": { "description": "Unsubscribed." }, "401": { "$ref": "#/components/responses/Unauthorized" } } }
    },
    "/push/device": {
      "post": { "tags": ["Push"], "summary": "Register a native APNs/FCM device token", "security": [ { "cookieAuth": [] } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["token"], "properties": { "token": { "type": "string" }, "platform": { "type": "string", "enum": ["ios", "android"] } } } } } }, "responses": { "204": { "description": "Registered." }, "401": { "$ref": "#/components/responses/Unauthorized" } } }
    },
    "/push/device/{token}": {
      "parameters": [ { "name": "token", "in": "path", "required": true, "schema": { "type": "string" } } ],
      "delete": { "tags": ["Push"], "summary": "Remove a device token", "security": [ { "cookieAuth": [] } ], "responses": { "204": { "description": "Removed." }, "401": { "$ref": "#/components/responses/Unauthorized" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "burncpu_session", "description": "Session cookie from the magic-link flow. HttpOnly; CSRF-protected." },
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "API token from POST /tokens. Scoped, non-session endpoints only." }
    },
    "parameters": {
      "IdPath": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
      "Limit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
      "Before": { "name": "before", "in": "query", "required": false, "description": "Pagination cursor — return items older than this.", "schema": { "type": "string" } }
    },
...
    "responses": {
      "Unauthorized": { "description": "Authentication required or invalid.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "Authenticated but not permitted.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Resource not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": {
        "type": "object", "required": ["error", "message"],
        "properties": {
          "error": { "type": "string", "enum": ["not_found", "unauthorized", "forbidden", "bad_request", "rate_limited", "internal"] },
          "message": { "type": "string" }
        }
      },
...
      "Author": {
        "type": "object", "required": ["id", "username", "display_name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "display_name": { "type": "string" },
          "avatar_url": { "type": "string" }
        }
      },
      "Post": {
        "type": "object", "required": ["id", "author", "body", "body_html", "visibility", "reactions_count", "replies_count", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "author": { "$ref": "#/components/schemas/Author" },
          "body": { "type": "string" },
          "body_html": { "type": "string", "description": "Sanitized HTML render of body." },
          "visibility": { "type": "string", "enum": ["public", "unlisted", "followers"] },
          "reply_to_id": { "type": ["string", "null"], "format": "uuid" },
          "parent": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" }, "author_username": { "type": "string" }, "excerpt": { "type": "string" } } },
          "content_warning": { "type": "string" },
          "reactions_count": { "type": "integer" },
          "replies_count": { "type": "integer" },
          "edited_at": { "type": "string", "format": "date-time", "description": "Present if the post has been edited; see /posts/{id}/history." },
          "viewer_reacted": { "type": "boolean", "description": "Present only for authenticated viewers." },
          "viewer_bookmarked": { "type": "boolean", "description": "Present only for authenticated viewers." },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": { "type": "string", "description": "e.g. follow, reaction, reply, mention, repost, dm" },
          "actor": { "$ref": "#/components/schemas/Author" },
          "post_id": { "type": ["string", "null"], "format": "uuid" },
          "read": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}